face_recognition/requirements.txt
Click==7.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.1
opencv-contrib-python-headless==4.2.0.32
opencv-python==3.3.1.11
Pillow==7.0.0
